suppress warnings
git-svn-id: https://svn.salilab.org/imp/trunk@16770 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/kernel/include/internal/ExponentialNumber.h
@@ -44,7 +44,7 @@ class ExponentialNumber
   void copy_from(ExponentialNumber<OEXP> o) {
     const int diff = OEXP-EXP;
     v_=o.v_;
-    float factor= std::pow(10.0, static_cast<double>(diff));
+    double factor= std::pow(10.0, static_cast<double>(diff));
     v_*=factor;
   }
   bool is_default() const {
